requirements.txt
evdev==1.7.1
pygame==2.5.2
RPi.GPIO==0.7.1
